Fixed a decorator that wasn't preserving function metadata correctly
